PageRank Update

+++ src/main/java/edu/brown/cs/termproject/model/User.java
@@ -79,7 +79,7 @@ public class User implements PageRankNode<Course> {
 
     ImmutableMap.Builder<Course, Double> builder = ImmutableMap.builder();
 
-    double weight = 1 / registrations.size();
+    double weight = 1 / (double) registrations.size();
     for (Registration registration : registrations) {
       builder.put(registration.getCourse(), weight);
     }
